Refactor scripts and clean GUI
